Hydro

Documentation for Hydro.

Hydro.Grid

Hydro.calc_flux!

Hydro.calc_flux!

Hydro.cons2prim

Hydro.cons2prim

Hydro.fill_periodic_bc

Hydro.fill_trans_bc

Hydro.fill_trans_bc

Hydro.hll1st

Hydro.hll1st

Hydro.hll2nd

Hydro.hll2nd

Hydro.hydro

Hydro.init_KH

Hydro.init_KH2

Hydro.init_ball

Hydro.init_sod

Hydro.init_sod

Hydro.plot_curve

Hydro.plot_curve_old

Hydro.plot_curve_or_heat

Hydro.plot_curve_or_heat

Hydro.plot_heat

Hydro.plot_heat_four_panels

Hydro.prim2cons!

Hydro.prim2cons!

Hydro.prim2cs

Hydro.reconstruct

Hydro.Grid — Type

Memory use: N * 5 # cs, E, epsilon, csL, csR 3N * 9 # u, fu, lu, fhll, uL, uR, prims, primsL, primsR tot memory = 32N

Hydro.calc_flux! — Method

2D: Calculate flux from prims (rho, vel, pressure)

Hydro.calc_flux! — Method

Calculate flux from prims (rho, vel, pressure)

Hydro.cons2prim — Method

2d: Reconstruct rho, vel, E, epsilon, and pressure from g.u

Hydro.cons2prim — Method

Reconstruct rho, vel, E, epsilon, and pressure from g.u

Hydro.fill_periodic_bc — Method

Fill periodic boundary conditions in 2D space

Hydro.fill_trans_bc — Method

Fill 2D boundaries: transimissive

Hydro.fill_trans_bc — Method

Fill 1D boundaries: transimissive

Hydro.hll1st — Method

HLL, 1st order. Based on Chapter 7.2.1 of Springel's class notes

Hydro.hll1st — Method

HLL Riemann solver, 1st order. Piecewise constant construction is applied where the left and right states on the interface are simply U{i+1/2},L = u{i} U{i+1/2},R = u{i+1} The wave-speeds are estimated via the simple expression: SL = min{UL - aL, UR - aR} SR = max{UL + aL, UR + aR}

Hydro.hll2nd — Method

Second-order HLL. This should not change g.u

Hydro.hll2nd — Method

Second-order HLL. This should not change g.u

Hydro.hydro — Method

A general-purpose hydro solver

Hydro.init_KH — Method

Kelvin-Helmholtz instability

Hydro.init_KH2 — Method

Kelvin-Helmholtz instability

Hydro.init_ball — Method

Set the 2d sod shock tube: a ball of radius 0.2 at center with density 1.0, pressure 1.0. Outside has density 0.125 and pressure 0.1

Hydro.init_sod — Method

Set the 2d initial conditions: a 2D version of the 1D sod shocktube

Hydro.init_sod — Method

set the initial conditions of a sod shock tube

Hydro.plot_curve — Method

Plot 1D curve from a 2d simulation

Hydro.plot_curve_old — Method

make the plot, 1D case

Hydro.plot_curve_or_heat — Method

Plot curve for a 1d simulation and heat for a 2d simulation

Hydro.plot_curve_or_heat — Method

Plot curve for a 1d simulation and heat for a 2d simulation

Hydro.plot_heat — Method

Plot heatmap of the density for a 2d simulation

Hydro.plot_heat_four_panels — Method

Plot heatmap of the density, 2D case

Hydro.prim2cons! — Method

2d: From primilary variables (rho, vel, pressure) to conservative variables

Hydro.prim2cons! — Method

From primilary variables (rho, vel, pressure) to conservative variables

Hydro.prim2cs — Method

From prims to sound speed

Hydro.reconstruct — Function

Interpolate prims (rhoL, rhoR, velL, velR, pressureL, pressureR), which are used to update the cons (uL, uR)