Hydro

Documentation for Hydro.

Hydro.Grid

Hydro.calc_flux!

Hydro.calc_flux!

Hydro.cons2prim

Hydro.cons2prim

Hydro.evolve

Hydro.fill_periodic_bc

Hydro.fill_trans_bc

Hydro.fill_trans_bc

Hydro.hll

Hydro.hll

Hydro.hllc

Hydro.hllc_var1

Hydro.hydro

Hydro.init_KH

Hydro.init_KH2

Hydro.init_ball

Hydro.init_sod

Hydro.init_sod

Hydro.init_sod_y

Hydro.plot_curve

Hydro.plot_curve_or_heat

Hydro.plot_curve_or_heat

Hydro.plot_heat_four_panels

Hydro.prim2cons!

Hydro.prim2cons!

Hydro.prim2cs

Hydro.reconstruct2nd

Hydro.reconstruct2nd

Hydro.Grid — Type

Memory use: N * 5 # cs, E, epsilon, csL, csR 3N * 9 # u, fu, lu, fhll, uL, uR, prims, primsL, primsR tot memory = 32N

Hydro.calc_flux! — Method

2D: Calculate flux from prims (rho, vel, pressure)

Hydro.calc_flux! — Method

Calculate flux from prims (rho, vel, pressure)

Hydro.cons2prim — Method

2d: Reconstruct rho, vel, E, epsilon, and pressure from g.u

Hydro.cons2prim — Method

Reconstruct rho, vel, E, epsilon, and pressure from g.u

Hydro.evolve — Method

evolve(s)

Run a hydrodynamics simulation using the parameters stored in the sim structure. This function evolves the simulation from the current time to the end time.

Hydro.fill_periodic_bc — Method

Fill periodic boundary conditions in 2D space

Hydro.fill_trans_bc — Method

Fill 2D boundaries: transimissive

Hydro.fill_trans_bc — Method

Fill 1D boundaries: transimissive

Hydro.hll — Function

Second-order HLL. This should not change g.u

Hydro.hll — Function

Second-order HLL.

Hydro.hllc — Function

HLLC scheme in 1D

Hydro.hllc_var1 — Function

HLLC solver. Work in progress. Some bugs remains to be fixed.

Hydro.hydro — Method

A general-purpose hydro solver

hydro(dim, nx, tend, folder, init; solver::Function=hll, order::Int=2, integrator::Function=RK3, fillbc::Function=filltransbc, plotit::Function=plotcurveor_heat, dtout::Float64=0.01, ny::Int64=-1, storealldata::Bool=false, restart=-1, islog::Bool=true, verbose::Bool=false)

Run a hydrodynamics simulation and save the figures in folder.

Arguments

dim: (Int) number of dimentions, either 1 or 2
nx: (Int) number of pixels in the first dimention (use ny= to set the number of pixels in the second dimension)
tend: (Float) the end time
folder: (String) the folder to store the outputs (figures, logs, and data)
init: (Function) a function to setup the initial condition. Choose one of init_sod, init_sod_y, init_ball, init_KH.
solver: (Function, default: hll) the Riemann solver. Choose one of hll, hllc, lax.
order: (Int, default: 2) 1 for first order accuracy in space and 2 for second order accuracy.
integrator: (Function, default: RK3) one of euler, RK3, RK3
fillbc: (Function) a function to set the boundary condition. One of fill_trans_bc, fill_periodic_bc.
plotit: (Function) plotting function. One of the following functions.
    plot_curve: plot curves of the density, velocity, pressure, and specific energy
    plot_heat: plot heat map of the density
    plot_heat_four_panels: plot heat maps of the density, x-velocity, y-velocity, and pressure
    plot_curve_or_heat (default): automatically pick plot_curve if dim=1 or plot_heat if dim=2
    plot_standard_sod: on top of plot_curve, plot the analytic solution to the standard Sod shock tube problem (rho = 1.0, 0.125, velocity = 0.0, 0.0, pressure = 1.0, 0.1 for the left and right state)
dtout: (Float, default: 0.01) the time interval between two outputs.
ny: (Int, default: -1) the number of pixels in the second dimention. A negative value (default) means ny = nx.
storealldata: (Bool, default: false) toggle store reloadable data for each outputs set by dtout. If false, will only store the data for the last output, or the last output before you interrupt with Ctrl-c.
restart: (Int, default: -1) the index of the output from which to restart the simulation. Default value -1 means starting from time 0. 
islog: (Bool, default: true) toggle creating a log file
verbose: (Bool, default: false) toggle verbose printing

Examples

julia> hydro(1, 512, 0.1, "tmp", init_sod)

julia> hydro(1, 512, 0.2, "tmp", init_sod, restart=10)

julia> hydro(2, 128, 1.0, "/tmp/Hydro/KH-128", init_KH; dtout=0.005, verbose=true)

Hydro.init_KH — Method

Kelvin-Helmholtz instability

Hydro.init_KH2 — Method

Kelvin-Helmholtz instability

Hydro.init_ball — Method

Set the 2d sod shock tube: a ball of radius 0.2 at center with density 1.0, pressure 1.0. Outside has density 0.125 and pressure 0.1

Hydro.init_sod — Method

Set the 2d initial conditions: a 2D version of the 1D sod shocktube

Hydro.init_sod — Method

set the initial conditions of a sod shock tube

Hydro.init_sod_y — Method

Set the 2d initial conditions: a 2D version of the 1D sod shocktube in the y dimension

Hydro.plot_curve — Method

Plot 1D curve from a 2d simulation

Hydro.plot_curve_or_heat — Method

Plot curve for a 1d simulation and heat for a 2d simulation

Hydro.plot_curve_or_heat — Method

Plot curve for a 1d simulation and heat for a 2d simulation

Hydro.plot_heat_four_panels — Method

Plot heatmap of the density, 2D case

Hydro.prim2cons! — Method

2d: From primilary variables (rho, vel, pressure) to conservative variables

Hydro.prim2cons! — Method

From primilary variables (rho, vel, pressure) to conservative variables

Hydro.prim2cs — Method

From prims to sound speed

Hydro.reconstruct2nd — Method

Interpolate primitive variables (ρ, vx, vy, p, cs) in the x and y components

Hydro.reconstruct2nd — Method

Interpolate prims (rhoL, rhoR, vxL, vxR, pressureL, pressureR), which are then used to update the cons (uL, uR)